new commit

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double weight_kg, double height_m) {
+    public int calculate(int weight_kg, double height_m) {
 
         double result;
         result = weight_kg / (height_m * height_m);
