+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
     public static void main(String[] args) {
         BmiService service = new BmiService();
-        double weight_kg = 98;
+        int weight_kg = 98;
         double height_m = 1.87;
